requirements.txt
asgiref==3.2.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-dropbox-storage==0.4.1
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
drf-yasg==1.17.1
dropbox==9.4.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
httpie==2.0.0
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.3
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
Pillow==7.1.2
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
beautifulsoup4==4.9.1
